requirements.txt
asgiref==3.11.0
Django==6.0
python-dateutil==2.9.0.post0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
wheel==0.45.1
